requirements.txt
# Core dependencies
fastapi==0.124.4
uvicorn[standard]==0.38.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1

